design: 알림 페이지 텍스트 왼쪽으로 수정, 네비게이션 알림 뱃지 색 하얀색으로 수정
